Removed code for SQLite
I have decided to use MySQL instead of SQLite. This is because MySQL will be more realistic for a game as users would be able to connect remotely to a database on a server (which would store their account and stats). The code will be easily modified so that this can happen. For now, the database will be stored locally.