Completed constructor of database class
Completed the constructor for the database class. It will create the database and tables (if they do not exist). Saved the database objects to class variables in the AccountSlot class. This means that the slot objects can query the database. This will be used to sign in accounts and get the account's data, etc.