Started GetUsernameAndPasswordWindow subroutine + additional minor changes
Started making the GetUsernameAndPasswordWindow subroutine. This will be used to allow a user to enter a username and password and will be used for signing into an account and for creating an account (as you will need a username and password for both cases). Additionally, I made some minor changes. For example, made the names of files more consistant.