Finished 'signOut' subroutine
'signOut' is used to sign an account out of a chosen account slot. This is done be setting all the data stored to none (other than the database object being used). This is so that another account can be signed into it. This will need to be tested once the 'signIn' subroutine is finished.