Completed 'viewAccount' subroutine
This subroutine (and associated subroutines) are used to allow the user to choose an account slot and have the stats for that account displayed to them (if an account is signed into the chosen slot). This will let the user see things such as number of wins (against the AI and against Players), etc.