Many changes
Added to GameAreaWindow.py, added to ChooseGameOptionsWindow.py, changed ButtonClass.py so that it doesn't draw the '3D border' if the colour of the button is black or white (since the 3D border uses black and white lines so it doesn't look right/effect doesn't work), grammar changes to 'into' vs 'in to' (might not still be correct). Progress: The user can now choose a game mode to play. They can choose the slot they wish to use if they are playing against the AI. If they chose to play against another player (as the game mode), both account slots will automatically be used and so they do not need to be given that choice. Checks are made to see if an account is signed in to the chosen slot (if PvAI) or slots (if PvP). Appropriate error messages are displayed to the user if the slot/s are empty.