Completed Functionality That Allows the User to Choose a Game Mode and the Options for That Game Mode
The user can choose betwen 'player vs player' (PvP) or 'player vs AI' (PvAI). If the user chooses PvP, they will (provided they meet the requirements to start a game, that being that an account is signed in to each slot) be able to choose the colour (black or white) for slot one (the colour not chosen will be given to slottwo automatically). If the user chooses PvAI, they will need to choose an account slot to use. If there is an account signed in to the chosen slot, they will then be able to choose the colour they want to play as (black or white) and the difficulty of the AI (four difficulty levels). The AI will automatically be play as the colour that the user didn't choose.