Completed 'processClick' subroutine (mostly)
'processClick' is now mostly completed it. It allows the user to select a man/king display its legal moves, select a different man/king (if they should be allowed to), and allows the user to unselect (and have no man/king selected) if they click an empty square which doesn't have a legal move circle (that square is not a destination of a legal move) (this again only happens if the user should be allowed to -> See comments in code. The uncompleted part of this subroutine is the selection statement which is used to carry out a move. Before this can be written (and therefore this subroutine fully complete) the subroutine that will be used to a move a man/king (carry out a legal move) must be written and completed. This will be done next.