Fixed ChooseGameOptionsWindows.py and Completed Code For Carrying out a move
ChooseGameOptionsWindows.py had a bug which meant that an error would be thrown when trying to start a PvP game. This has been fixed by changing the selection statements in the event loop. Code for carrying out a move has been completed by finishing the 'processClick' and 'endTurn' subroutines. A bug has been found which means that when the turn switches to COLOUR_TWO, it is not finding any legal moves. This needs to be fixed next.