Added Code to Check for if Someone Has Won, if There Is a Draw or if the Game Should Continue + Bug Fixes
Added the code/subroutines to check if the game should end or if it should continue. A bug was found which meant that if a man/king got promoted by capturing an opponent man/king, when discovering the legal moves it could keep jumping back and forth over that same opponent man/king as technically the man/king was still on the board. This meant it would keep discovering the same moves infinitely. This has been fixed by storing each move discovered and checking if a move just discovered would be doing the reverse of a previous move. This stops it going back over the same man/king as it a move like that is found, it is not added to the tree or queue.