Bug Fixes
Fixed bugs that sometimes crashed the game when moving a man/king. This was linked to the fact legal moves were still being found when a man was promoted to a king. There was a mistake on my part as I had forgotten to make use of the 'isAlreadyKing' variable. Everything seems to work fine now. In the future I will need to work on redoing large amounts of the project since it is starting to get quite messy and have lots of repeated and inefficent code. I do not have long left until my coursework is due so i want to focus on getting the project done first.