Started Minimax.py, completed scoreBoard() + cleanup
I have started Minimax.py which will store some of the subroutines that are needed to carry out a AI's move (AI will use minimax). I have completed the subroutine scoreBoard() which is used to give a board a score. This score is needed for minimax() to work out which moves are best. I have also removed the print statements from the code that were there when i was fixing previous bugs. They should all be removed now.