Completed getLegalMovesAsList() and getRootToLeafPaths()
getLegalMovesAsList() will make a tree of legal moves and then use getRootToLeafPaths() to get a list of all root-to-leaf paths of the tree. Each path will represent a turn that a player could make (one path can have multiple moves since the player might make 2+ captures in a single turn. This path would show all the moves that would make up that turn.)