Updated workOutIsGameFinished()
Update workOutIsGameFinished() to improve and fix the logic used to decide if a game is over. I am considering moving it to the Board class so that the same subroutine can be used for the AI/minimax() subroutine in the Minimax.py file to decide if the recursive subroutine should start unwinding. This is because currently minimax() only checks the depth and if the current deepcopy board has a colour (COLOUR_ONE or COLOUR_TWO) with no men/kings left.