Program Complete - Multiple Changes
The whole program should now be complete. It will require most testing but everything I have tried so far is working. I have fixed some of the logic for modifying a users stats at the end of the game because some stats were missing. They should all be working now. I have also moved workOutIsGameFinished() to the Board class so that the same method can be used by the minimax() algorithm when it deepcopies a board. workOutIsGameFinished() used to be in the Game class but this meant that only the game object (of which there is only ever one at a time). Now that it is in the Board (which it is better suited in regardless) it means that the minimax() algorithm that it used for the AI's turn can make use of it to determine when it should unwind.