Fixed AI Depths + Tidied Code
I have adjusted the difficulty depths used for the AI and tested them and i think that better reflect the difficulty choices and makes sure that it is a good balance of difficulty and wait times (for the minimax algoritm to work out the best move). I have also gone through every file to make sure they follow the same layout in the files, added more commentes, and made everything more consistent.